+++ app/controllers/projects_controller.rb
@@ -71,7 +71,7 @@ class ProjectsController < ApplicationController
   # DELETE /projects/1
   def destroy
     @project.destroy
-    redirect_to projects_path, notice: 'Project was successfully destroyed.'
+    redirect_to projects_path, notice: 'Project was successfully deleted.'
   end
 
   private