+++ app/controllers/users_controller.rb
@@ -43,7 +43,7 @@ class UsersController < ApplicationController
   def destroy
     @user.destroy
     respond_to do |format|
-      format.html { redirect_to users_path, notice: 'User was successfully destroyed.' }
+      format.html { redirect_to users_path, notice: 'User was successfully deleted.' }
       format.json { head :no_content }
     end
   end