+++ app/models/document.rb
@@ -3,13 +3,14 @@
 # Allows documents to be attached to categories.
 class Document < ApplicationRecord
   # Uploaders
+  mount_uploader :primary_document, PrimaryDocumentUploader
   mount_uploader :document, DocumentUploader
 
   # Concerns
   include Deletable
 
   # Model Validation
-  validates :project_id, :user_id, :category_id, :document, presence: true
+  validates :project_id, :user_id, :category_id, :primary_document, presence: true
 
   # Model Relationships
   belongs_to :project
@@ -19,14 +20,20 @@ class Document < ApplicationRecord
   # Document Methods
 
   def name
+    primary_document_identifier || document_identifier
+  end
+
+  def name_secondary
     document_identifier
   end
 
   def pdf?
-    document_identifier.last(4).to_s.downcase == '.pdf'
+    return false unless primary_document.present?
+    primary_document_identifier.last(4).to_s.downcase == '.pdf'
   end
 
   def image?
-    document_identifier.last(4).to_s.downcase == '.png'
+    return false unless primary_document.present?
+    primary_document_identifier.last(4).to_s.downcase == '.png'
   end
 end