+++ db/migrate/20161010174653_add_primary_document_to_documents.rb
@@ -1,0 +1,5 @@
+class AddPrimaryDocumentToDocuments < ActiveRecord::Migration[5.0]
+  def change
+    add_column :documents, :primary_document, :string
+  end
+end